Add the User Additions module to the QB64 compiler build at setup time.

+++ setup_osx.command
@@ -44,6 +44,17 @@ if [ ! -f ./src.o ]; then
 fi
 cd ../../../../../../../..
 
+echo "Building User Additions"
+cd internal/c/parts/user_mods/os/lnx
+rm -f src.a
+./setup_build.sh
+if [ ! -f ./src.a ]; then
+  echo "Compilation of ./internal/c/parts/user_mods/os/osx/src.a failed!"
+  Pause
+  exit 1
+fi
+cd ../../../../../..
+
 echo "Building 'QB64' (~3 min)"
 cp ./internal/source/* ./internal/temp/
 cd internal/c
